banking_admin/requirements.txt
django>=5.1.4
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.0
